implement delete and done
